chore(CookieBanner): move out of SkyraSurvey into its own component
