+++ lib/components/CookieBanner/CookieBanner.tsx
