+++ lib/components/CookieBanner/__screenshots__/CookieBanner/CustomText.png
