+++ lib/components/CookieBanner/__screenshots__/CookieBanner/Default.png
