+++ lib/components/CookieBanner/__screenshots__/CookieBanner/Nynorsk.png
