+++ lib/components/CookieBanner/cookieBanner.module.css
@@ -1,7 +1,7 @@
 .banner {
   background-color: #0b4172;
   color: #ffffff;
-  padding: var(--ds-size-6) var(--dsc-spacing-page, var(--ds-size-6));
+  padding: 1.5rem;
   font-size: var(--ds-body-md-font-size);
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
