+++ lib/components/SkyraSurvey/index.ts
@@ -1,3 +1,2 @@
 export * from './SkyraSurvey';
-export * from './CookieBanner';
 export * from './useSkyraReload';
